requirements.txt
markitdown>=0.0.1
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
mcp>=0.9.0
python-docx>=1.0.0
pypdf>=3.17.0
Pillow>=10.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
python-pptx>=0.6.21
fastapi>=0.104.0
uvicorn>=0.24.0
sse-starlette>=1.6.5
fastmcp>=0.2.0
